requirements.txt
python-telegram-bot>=20.0
APScheduler==3.10.4
supabase>=1.0.0
python-dotenv>=0.19.0